refactor: extract CLI provisioning services
